HAL_ChibiOS: moved support for MCUs to their own module
this will make it easier to add more MCU types
