+++ libraries/AP_HAL_ChibiOS/hwdef/scripts/dma_resolver.py
@@ -1,179 +1,7 @@
 #!/usr/bin/env python
 
 import sys, fnmatch
-
-STM32F427_DMA_Map = {
-	# format is [DMA_TABLE, StreamNum]
-	# extracted from tabula-STM324x7-306.csv and tabula-STM324x7-307.csv
-	"ADC1"		:	[(2,0),(2,4)],
-	"ADC2"		:	[(2,2),(2,3)],
-	"ADC3"		:	[(2,0),(2,1)],
-	"CRYP_IN"	:	[(2,6)],
-	"CRYP_OUT"	:	[(2,5)],
-	"DAC1"		:	[(1,5)],
-	"DAC2"		:	[(1,6)],
-	"DCMI"		:	[(2,1),(2,7)],
-	"HASH_IN"	:	[(2,7)],
-	"I2C1_RX"	:	[(1,0),(1,5)],
-	"I2C1_TX"	:	[(1,6),(1,7)],
-	"I2C2_RX"	:	[(1,2),(1,3)],
-	"I2C2_TX"	:	[(1,7)],
-	"I2C3_RX"	:	[(1,2)],
-	"I2C3_TX"	:	[(1,4)],
-	"I2S2_EXT_RX"	:	[(1,3)],
-	"I2S2_EXT_TX"	:	[(1,4)],
-	"I2S3_EXT_RX"	:	[(1,2),(1,0)],
-	"I2S3_EXT_TX"	:	[(1,5)],
-	"SAI1_A"	:	[(2,1),(2,3)],
-	"SAI1_B"	:	[(2,5),(2,4)],
-	"SDIO"		:	[(2,3),(2,6)],
-	"SPI1_RX"	:	[(2,0),(2,2)],
-	"SPI1_TX"	:	[(2,3),(2,5)],
-	"SPI2_RX"	:	[(1,3)],
-	"SPI2_TX"	:	[(1,4)],
-	"SPI3_RX"	:	[(1,0),(1,2)],
-	"SPI3_TX"	:	[(1,5),(1,7)],
-	"SPI4_RX"	:	[(2,0),(2,3)],
-	"SPI4_TX"	:	[(2,1),(2,4)],
-	"SPI5_RX"	:	[(2,3),(2,5)],
-	"SPI5_TX"	:	[(2,4),(2,6)],
-	"SPI6_RX"	:	[(2,6)],
-	"SPI6_TX"	:	[(2,5)],
-	"TIM1_CH1"	:	[(2,6),(2,1),(2,3)],
-	"TIM1_CH2"	:	[(2,6),(2,2)],
-	"TIM1_CH3"	:	[(2,6),(2,6)],
-	"TIM1_CH4"	:	[(2,4)],
-	"TIM1_COM"	:	[(2,4)],
-	"TIM1_TRIG"	:	[(2,0),(2,4)],
-	"TIM1_UP"	:	[(2,5)],
-	"TIM2_CH1"	:	[(1,5)],
-	"TIM2_CH2"	:	[(1,6)],
-	"TIM2_CH3"	:	[(1,1)],
-	"TIM2_CH4"	:	[(1,6),(1,7)],
-	"TIM2_UP"	:	[(1,1),(1,7)],
-	"TIM3_CH1"	:	[(1,4)],
-	"TIM3_CH2"	:	[(1,5)],
-	"TIM3_CH3"	:	[(1,7)],
-	"TIM3_CH4"	:	[(1,2)],
-	"TIM3_TRIG"	:	[(1,4)],
-	"TIM3_UP"	:	[(1,2)],
-	"TIM4_CH1"	:	[(1,0)],
-	"TIM4_CH2"	:	[(1,3)],
-	"TIM4_CH3"	:	[(1,7)],
-	"TIM4_UP"	:	[(1,6)],
-	"TIM5_CH1"	:	[(1,2)],
-	"TIM5_CH2"	:	[(1,4)],
-	"TIM5_CH3"	:	[(1,0)],
-	"TIM5_CH4"	:	[(1,1),(1,3)],
-	"TIM5_TRIG"	:	[(1,1),(1,3)],
-	"TIM5_UP"	:	[(1,0),(1,6)],
-	"TIM6_UP"	:	[(1,1)],
-	"TIM7_UP"	:	[(1,2),(1,4)],
-	"TIM8_CH1"	:	[(2,2),(2,2)],
-	"TIM8_CH2"	:	[(2,2),(2,3)],
-	"TIM8_CH3"	:	[(2,2),(2,4)],
-	"TIM8_CH4"	:	[(2,7)],
-	"TIM8_COM"	:	[(2,7)],
-	"TIM8_TRIG"	:	[(2,7)],
-	"TIM8_UP"	:	[(2,1)],
-	"UART4_RX"	:	[(1,2)],
-	"UART4_TX"	:	[(1,4)],
-	"UART5_RX"	:	[(1,0)],
-	"UART5_TX"	:	[(1,7)],
-	"UART7_RX"	:	[(1,3)],
-	"UART7_TX"	:	[(1,1)],
-	"UART8_RX"	:	[(1,6)],
-	"UART8_TX"	:	[(1,0)],
-	"USART1_RX"	:	[(2,2),(2,5)],
-	"USART1_TX"	:	[(2,7)],
-	"USART2_RX"	:	[(1,5)],
-	"USART2_TX"	:	[(1,6)],
-	"USART3_RX"	:	[(1,1)],
-	"USART3_TX"	:	[(1,3),(1,4)],
-	"USART6_RX"	:	[(2,1),(2,2)],
-	"USART6_TX"	:	[(2,6),(2,7)],
-}
-
-STM32F412_DMA_Map = {
-	# format is [DMA_TABLE, StreamNum]
-	# extracted from tabula-stm32f412-ref-manual-196.csv and tabula-stm32f412-ref-manual-196(1).csv
-	"ADC1"		:	[(2,0),(2,4)],
-	"DFSDM1_FLT0"	:	[(2,6),(2,0)],
-	"DFSDM1_FLT1"	:	[(2,1),(2,4)],
-	"I2C1_RX"	:	[(1,0),(1,5)],
-	"I2C1_TX"	:	[(1,1),(1,6),(1,7)],
-	"I2C2_RX"	:	[(1,2),(1,3)],
-	"I2C2_TX"	:	[(1,7)],
-	"I2C3_RX"	:	[(1,1),(1,2)],
-	"I2C3_TX"	:	[(1,4),(1,5)],
-	"I2CFMP1_RX"	:	[(1,3),(1,0)],
-	"I2CFMP1_TX"	:	[(1,1),(1,7)],
-	"I2S2EXT_RX"	:	[(1,3)],
-	"I2S2_EXT_TX"	:	[(1,4)],
-	"I2S3_EXT_RX"	:	[(1,2),(1,0)],
-	"I2S3_EXT_TX"	:	[(1,5)],
-	"QUADSPI"	:	[(2,7)],
-	"SDIO"		:	[(2,3),(2,6)],
-	"SPI1_RX"	:	[(2,0),(2,2)],
-	"SPI1_TX"	:	[(2,2),(2,3),(2,5)],
-	"SPI2_RX"	:	[(1,3)],
-	"SPI2_TX"	:	[(1,4)],
-	"SPI3_RX"	:	[(1,0),(1,2)],
-	"SPI3_TX"	:	[(1,5),(1,7)],
-	"SPI4_RX"	:	[(2,0),(2,4),(2,3)],
-	"SPI4_TX"	:	[(2,1),(2,4)],
-	"SPI5_RX"	:	[(2,3),(2,5)],
-	"SPI5_TX"	:	[(2,4),(2,5),(2,6)],
-	"TIM1_CH1"	:	[(2,6),(2,1),(2,3)],
-	"TIM1_CH2"	:	[(2,6),(2,2)],
-	"TIM1_CH3"	:	[(2,6),(2,6)],
-	"TIM1_CH4"	:	[(2,4)],
-	"TIM1_COM"	:	[(2,4)],
-	"TIM1_TRIG"	:	[(2,0),(2,4)],
-	"TIM1_UP"	:	[(2,5)],
-	"TIM2_CH1"	:	[(1,5)],
-	"TIM2_CH2"	:	[(1,6)],
-	"TIM2_CH3"	:	[(1,1)],
-	"TIM2_CH4"	:	[(1,6),(1,7)],
-	"TIM2_UP"	:	[(1,1),(1,7)],
-	"TIM3_CH1"	:	[(1,4)],
-	"TIM3_CH2"	:	[(1,5)],
-	"TIM3_CH3"	:	[(1,7)],
-	"TIM3_CH4"	:	[(1,2)],
-	"TIM3_TRIG"	:	[(1,4)],
-	"TIM3_UP"	:	[(1,2)],
-	"TIM4_CH1"	:	[(1,0)],
-	"TIM4_CH2"	:	[(1,3)],
-	"TIM4_CH3"	:	[(1,7)],
-	"TIM4_UP"	:	[(1,6)],
-	"TIM5_CH1"	:	[(1,2)],
-	"TIM5_CH2"	:	[(1,4)],
-	"TIM5_CH3"	:	[(1,0)],
-	"TIM5_CH4"	:	[(1,1),(1,3)],
-	"TIM5_TRIG"	:	[(1,1),(1,3)],
-	"TIM5_UP"	:	[(1,0),(1,6)],
-	"TIM6_UP"	:	[(1,1)],
-	"TIM7_UP"	:	[(1,2),(1,4)],
-	"TIM8_CH1"	:	[(2,2),(2,2)],
-	"TIM8_CH2"	:	[(2,2),(2,3)],
-	"TIM8_CH3"	:	[(2,2),(2,4)],
-	"TIM8_CH4"	:	[(2,7)],
-	"TIM8_COM"	:	[(2,7)],
-	"TIM8_TRIG"	:	[(2,7)],
-	"TIM8_UP"	:	[(2,1)],
-	"USART1_RX"	:	[(2,2),(2,5)],
-	"USART1_TX"	:	[(2,7)],
-	"USART2_RX"	:	[(1,5),(1,7)],
-	"USART2_TX"	:	[(1,6)],
-	"USART3_RX"	:	[(1,1)],
-	"USART3_TX"	:	[(1,3),(1,4)],
-	"USART6_RX"	:	[(2,1),(2,2)],
-	"USART6_TX"	:	[(2,6),(2,7)],
-}
-
-# map above dma maps to MCU types. The MCU type strings must match those used in ChibiOS
-dma_maps = { "STM32F412Rx" : STM32F412_DMA_Map,
-             "STM32F427xx" : STM32F427_DMA_Map }
+import importlib
 
 # peripheral types that can be shared, wildcard patterns
 SHARED_MAP = [ "I2C*", "USART*_TX", "UART*_TX", "SPI*" ]
@@ -239,21 +67,25 @@ def chibios_dma_define_name(key):
 
 def write_dma_header(outfilename, peripheral_list, mcu_type):
         '''write out a DMA resolver header file'''
+        global dma_map
+        try:
+                lib = importlib.import_module(mcu_type)
+                dma_map = lib.DMA_Map
+        except ImportError:
+                print("Unable to find module for MCU %s" % mcu_type)
+                sys.exit(1)
+        
         print("Writing DMA map to %s" % outfilename)
         f = open(outfilename, 'w')
         unassigned = []
         curr_dict = {}
 
-        if not mcu_type in dma_maps:
-                print("No DMA map for MCU %s" % mcu_type)
-                sys.exit(1)
-
-        global dma_map
-        dma_map = dma_maps[mcu_type]
-
 	for periph in peripheral_list:
 		assigned = False
 		check_list = []
+                if not periph in dma_map:
+                        print("Unknown peripheral function %s in DMA map for %s" % (periph, mcu_type))
+                        sys.exit(1)
 		for stream in dma_map[periph]:
 			if check_possibility(periph, stream, curr_dict, dma_map, check_list):
 				curr_dict[periph] = stream
@@ -329,40 +161,24 @@ def write_dma_header(outfilename, peripheral_list, mcu_type):
 if __name__ == '__main__':
 	import optparse
 
-	boards = ['F412', 'FMUv3']
-
 	parser = optparse.OptionParser("dma_resolver.py")
-	parser.add_option("-B", "--board", type='choice', default=None, choices=boards, help='board type')
+	parser.add_option("-M", "--mcu", default=None, help='MCU type')
 	parser.add_option("-D", "--debug", action='store_true', help='enable debug')
+	parser.add_option("-P", "--peripherals", default=None, help='peripheral list (comma separated)')
 
 	opts, args = parser.parse_args()
 
-	if opts.board == 'FMUv3':
-                mcu_type = "STM32F427"
-	        PERIPHONDMA_LIST = ["SDIO"]
-	        PERIPHONDMA_LIST += ["ADC1"]
-	        PERIPHONDMA_LIST += ["SPI1_RX","SPI1_TX","SPI2_RX","SPI2_TX","SPI4_RX","SPI4_TX"]
-	        PERIPHONDMA_LIST += ["I2C1_RX","I2C1_TX","I2C2_RX","I2C2_TX"]
-	        PERIPHONDMA_LIST += ["USART6_TX","USART6_RX"] # px4io
-	        PERIPHONDMA_LIST += ["USART2_TX","USART2_RX"] # telem1
-	        PERIPHONDMA_LIST += ["USART3_TX","USART3_RX"] # telem2
-	        PERIPHONDMA_LIST += ["UART4_TX","UART4_RX"]   # GPS
-	        PERIPHONDMA_LIST += ["UART8_TX","UART8_RX"]   # GPS2
-	        PERIPHONDMA_LIST += ["UART7_TX","UART7_RX"]   # console
-	        PERIPHONDMA_LIST += ["USART1_TX","USART1_RX"] # not used
-	elif opts.board == 'F412':
-                mcu_type = "STM32F412"
-	        PERIPHONDMA_LIST = ["ADC1"]
-	        PERIPHONDMA_LIST += ["SPI1_RX","SPI1_TX","SPI2_RX","SPI2_TX"]
-	        PERIPHONDMA_LIST += ["I2C1_RX","I2C1_TX","I2C2_RX", "I2C2_TX"]
-	        PERIPHONDMA_LIST += ["USART2_TX","USART2_RX"]
-	        PERIPHONDMA_LIST += ["USART6_TX","USART6_RX"]
-	        PERIPHONDMA_LIST += ["USART1_TX","USART1_RX"]
+        if opts.peripherals is None:
+                print("Please provide a peripheral list with -P")
+                sys.exit(1)
 
-	if opts.board is None:
-	        print("Please choose a board from: %s" % boards)
-	        sys.exit(1)
+        if opts.mcu is None:
+                print("Please provide a MCU type with -<")
+                sys.exit(1)
 
         debug = opts.debug
 
-        write_dma_header("dma.h", PERIPHONDMA_LIST, mcu_type)
+        plist = opts.peripherals.split(',')
+        mcu_type = opts.mcu
+
+        write_dma_header("dma.h", plist, mcu_type)
